Link screens and fix pages
