+++ .github/workflows/test.yml
@@ -18,5 +18,6 @@ jobs:
           npm install
           npm --prefix backend install
           npm --prefix frontend install
+          npm install jest-environment-jsdom
       - name: Run tests
         run: npm test
