+++ backend/jest.config.js
@@ -1,6 +1,18 @@
 module.exports = {
   preset: 'ts-jest',
+  globals: {
+    'ts-jest': {
+      tsconfig: 'tsconfig.spec.json',
+    },
+  },
+  moduleFileExtensions: ['js', 'json', 'ts'],
+  testPathIgnorePatterns: ['/node_modules/', '<rootDir>/node_modules/'],
+  rootDir: 'src',
+  testRegex: '(/__tests__/.*|\\.(test|spec))\\.(ts|js)$',
+  transform: {
+    '^.+\\.ts$': 'ts-jest',
+  },
+  collectCoverageFrom: ['**/*.(t|j)s'],
+  coverageDirectory: './coverage',
   testEnvironment: 'node',
-  roots: ['<rootDir>/src'],
-  testMatch: ['**/?(*.)+(spec|test).ts'],
 };
