+++ backend/package.json
@@ -33,39 +33,11 @@
     "express": "^4.18.2"
   },
   "devDependencies": {
-
     "jest": "^29.0.0",
     "ts-jest": "^29.0.0",
     "@types/jest": "^29.0.0",
     "typescript": "^5.0.0",
     "cross-env": "^5.2.1",
     "rimraf": "^5.0.0"
-
-  },
-  "jest": {
-    "globals": {
-      "ts-jest": {
-        "tsconfig": "tsconfig.spec.json"
-      }
-    },
-    "moduleFileExtensions": [
-      "js",
-      "json",
-      "ts"
-    ],
-    "testPathIgnorePatterns": [
-      "/node_modules/",
-      "<rootDir>/node_modules/"
-    ],
-    "rootDir": "src",
-    "testRegex": "(/__tests__/.*|\\.(test|spec))\\.(ts|js)$",
-    "transform": {
-      ".ts": "ts-jest"
-    },
-    "collectCoverageFrom": [
-      "**/*.(t|j)s"
-    ],
-    "coverageDirectory": "./coverage",
-    "testEnvironment": "node"
   }
 }
